AdMob gone, New Donations View, refactors.
AboutFragment used to have a DonateButtonListener and an extension of
some sort of Purchase Listener.

I refactored all that logic into a Biller class, that I can use inside
AboutFragment and elsewhere.

However, I can't test because damn android won't create a resource id
for my new view. Looking for issues on the new XML to see why this is
happening.