Don't let the user download .torrents if they have said so on their Network Preferences.
If the user is on a mobile connection (3G/4G), a few seconds after the transfer starts, it will go to Queue state and a long toast will be shown letting the user know what's going on.
In the event the torrent transfer starts on a mobile connection and the preference is to not care about data consumption, a warning message and tip to check for Preferences is shown. This should address some complains from users that can't seem to tell when they're on Wi-Fi vs 3G/4G.

Needs a bit more testing, specially on the part where the Queue transfer is supposed to start automatically when the Wi-Fi network is available.